Use double instead of float to represent positions
This allow two things:
    - Represent exactly data from trajectories formats that uses
      doubles;
    - Gain some precision for analysis algorithms, where using float may
      lead to imprecision.

The memory usage should go up, but this should not really be a problem
most of the time.

+++ include/chemfiles.h
@@ -78,7 +78,7 @@ typedef enum {
     CHFL_CXX_ERROR = 6
 } chfl_status;
 
-typedef float chfl_vector_t[3];
+typedef double chfl_vector_t[3];
 
 /*!
 * @brief Get the version of the chemfiles library
